A classic Snake Game
Use of HTML, CSS and Javascript to make a classic Snake Game. The snake has to keep eating the apples to grow its length. The snake can neither touch any walls nor itself, otherwise the game will be over and have to start from the start. I use JavaScript to do that. Enjoy the game